+++ openless-all/app/src-tauri/src/persistence/credentials.rs
@@ -1,0 +1,880 @@
+#![cfg_attr(target_os = "linux", allow(dead_code, unused_variables))]
+//! Credentials vault.
+//!
+//! 正常读写走系统凭据库；旧 plaintext JSON 只作为迁移来源。为保持多 provider
+//! schema 与 active provider 状态，凭据库里保存一个 v1 JSON payload；payload 会按平台
+//! 凭据库限制拆成多个条目，避免 Windows 单条凭据 2560 bytes 限制。
+//!
+//! v1 schema：
+//!   {
+//!     "version": 1,
+//!     "active": { "asr": "<id>", "llm": "<id>" },
+//!     "providers": {
+//!       "asr": { "<id>": { "appKey", "accessKey", "resourceId", "apiKey", "baseURL", "model", "vocabularyId" } },
+//!       "llm": { "<id>": { "displayName", "apiKey", "baseURL", "model", "temperature", "extraHeaders" } }
+//!     }
+//!   }
+//!
+//! "ark.api_key"/"volcengine.app_key" 等账户名按 Swift 语义路由到 active provider。
+
+use std::collections::HashMap;
+use std::path::{Path, PathBuf};
+use std::sync::OnceLock;
+
+use anyhow::{Context, Result};
+use parking_lot::Mutex;
+use serde::{Deserialize, Serialize};
+
+// `anyhow!` is only invoked from the keyring (non-Android) code paths; gating the
+// import keeps the Android build free of an unused-import warning.
+#[cfg(not(target_os = "android"))]
+use anyhow::anyhow;
+#[cfg(target_os = "android")]
+use std::fs;
+
+/// 旧版 plaintext JSON 凭据路径。仅作为迁移来源；成功写入系统凭据库后会删除。
+const LEGACY_CREDS_DIR: &str = ".openless";
+const LEGACY_CREDS_FILE: &str = "credentials.json";
+
+const KEYRING_CREDENTIALS_ACCOUNT: &str = "credentials.v1";
+const KEYRING_CREDENTIALS_CHUNK_PREFIX: &str = "credentials.v1.chunk.";
+#[cfg(target_os = "android")]
+const ANDROID_CREDENTIALS_FILE: &str = "credentials.enc.json";
+// Windows Credential Manager caps one credential blob at 2560 bytes. keyring stores
+// passwords as UTF-16 on Windows, so keep each JSON chunk comfortably below that.
+const KEYRING_CHUNK_MAX_UTF16_UNITS: usize = 1000;
+
+static CREDENTIALS_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
+
+fn credentials_lock() -> &'static Mutex<()> {
+    CREDENTIALS_LOCK.get_or_init(|| Mutex::new(()))
+}
+
+/// Process-wide credentials cache.
+///
+/// Without this cache every `CredentialsVault::get_*` / `snapshot` call hits
+/// `load_credentials()` → `load_keyring_credentials()` which reads the
+/// manifest entry plus every chunk entry from the OS keyring. On macOS each
+/// distinct keychain entry has its own ACL — so an ad-hoc-signed binary (or
+/// any binary whose ACL grants haven't been set up yet) prompts on every read
+/// of every entry. A single dictation cycle reads credentials 5–10 times,
+/// times (1 manifest + N chunks) entries → tens of "OpenLess wants to use
+/// the keychain" prompts per recording.
+///
+/// With this cache the first read populates `Some(CredsRoot)` and every
+/// subsequent read in the same process is silent. `save_credentials` keeps
+/// the cache in sync after writes so Settings → Recording credential edits
+/// take effect immediately.
+///
+/// Cross-process changes (e.g. user edits via `security` CLI, or another
+/// instance of the app — single-instance is enforced but defense in depth)
+/// will be invisible until the next process launch. Acceptable trade-off
+/// per the credential vault contract: the keyring is owned by this app.
+static CREDENTIALS_CACHE: OnceLock<Mutex<Option<CredsRoot>>> = OnceLock::new();
+
+fn credentials_cache() -> &'static Mutex<Option<CredsRoot>> {
+    CREDENTIALS_CACHE.get_or_init(|| Mutex::new(None))
+}
+
+fn store_credentials_cache(root: &CredsRoot) {
+    *credentials_cache().lock() = Some(root.clone());
+}
+
+#[cfg(test)]
+fn reset_credentials_cache_for_tests() {
+    *credentials_cache().lock() = None;
+}
+
+#[derive(Debug, Serialize, Deserialize, Default, Clone)]
+#[allow(non_snake_case)]
+struct CredsRoot {
+    #[serde(default = "credsroot_default_version")]
+    version: u32,
+    #[serde(default)]
+    active: CredsActive,
+    #[serde(default)]
+    providers: CredsProviders,
+}
+
+fn credsroot_default_version() -> u32 {
+    1
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+struct CredsActive {
+    #[serde(default = "creds_default_asr")]
+    asr: String,
+    #[serde(default = "creds_default_llm")]
+    llm: String,
+}
+
+impl Default for CredsActive {
+    fn default() -> Self {
+        Self {
+            asr: creds_default_asr(),
+            llm: creds_default_llm(),
+        }
+    }
+}
+
+fn creds_default_asr() -> String {
+    #[cfg(target_os = "windows")]
+    {
+        return crate::asr::local::foundry::PROVIDER_ID.into();
+    }
+    #[cfg(not(target_os = "windows"))]
+    {
+        "volcengine".into()
+    }
+}
+fn creds_default_llm() -> String {
+    "ark".into()
+}
+
+#[derive(Debug, Serialize, Deserialize, Default, Clone)]
+struct CredsProviders {
+    #[serde(default)]
+    asr: HashMap<String, CredsAsrEntry>,
+    #[serde(default)]
+    llm: HashMap<String, CredsLlmEntry>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Default, Clone)]
+#[allow(non_snake_case)]
+struct CredsAsrEntry {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    apiKey: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    baseURL: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    model: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    appKey: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    accessKey: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    resourceId: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    vocabularyId: Option<String>,
+}
+
+impl CredsAsrEntry {
+    fn is_empty(&self) -> bool {
+        self.apiKey.as_deref().unwrap_or("").is_empty()
+            && self.baseURL.as_deref().unwrap_or("").is_empty()
+            && self.model.as_deref().unwrap_or("").is_empty()
+            && self.appKey.as_deref().unwrap_or("").is_empty()
+            && self.accessKey.as_deref().unwrap_or("").is_empty()
+            && self.resourceId.as_deref().unwrap_or("").is_empty()
+            && self.vocabularyId.as_deref().unwrap_or("").is_empty()
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, Default, Clone)]
+#[allow(non_snake_case)]
+struct CredsLlmEntry {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    displayName: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    apiKey: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    baseURL: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    model: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    temperature: Option<f64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    extraHeaders: Option<HashMap<String, String>>,
+}
+
+impl CredsLlmEntry {
+    fn is_empty(&self) -> bool {
+        self.displayName.as_deref().unwrap_or("").is_empty()
+            && self.apiKey.as_deref().unwrap_or("").is_empty()
+            && self.baseURL.as_deref().unwrap_or("").is_empty()
+            && self.model.as_deref().unwrap_or("").is_empty()
+            && self.temperature.is_none()
+            && self
+                .extraHeaders
+                .as_ref()
+                .map(|h| h.is_empty())
+                .unwrap_or(true)
+    }
+}
+
+fn credentials_path() -> Result<PathBuf> {
+    // macOS / Linux: ~/.openless/credentials.json (与 Swift 同源)
+    // Windows: %APPDATA%\OpenLess\credentials.json (Windows 没有标准 HOME 环境变量)
+    #[cfg(target_os = "windows")]
+    {
+        let appdata = std::env::var("APPDATA").context("APPDATA not set")?;
+        return Ok(PathBuf::from(appdata)
+            .join("OpenLess")
+            .join(LEGACY_CREDS_FILE));
+    }
+    #[cfg(not(target_os = "windows"))]
+    {
+        let home = std::env::var("HOME").context("HOME not set")?;
+        Ok(PathBuf::from(home)
+            .join(LEGACY_CREDS_DIR)
+            .join(LEGACY_CREDS_FILE))
+    }
+}
+
+#[cfg(not(target_os = "android"))]
+fn keyring_entry() -> Result<keyring::Entry> {
+    keyring_entry_for(KEYRING_CREDENTIALS_ACCOUNT)
+}
+
+#[cfg(not(target_os = "android"))]
+fn keyring_entry_for(account: &str) -> Result<keyring::Entry> {
+    keyring::Entry::new(CredentialsVault::SERVICE_NAME, account)
+        .context("open system credential vault")
+}
+
+#[cfg(target_os = "android")]
+fn android_credentials_path() -> Result<PathBuf> {
+    Ok(super::data_dir()?.join(ANDROID_CREDENTIALS_FILE))
+}
+
+#[cfg(target_os = "android")]
+fn load_android_credentials() -> Result<Option<CredsRoot>> {
+    let path = android_credentials_path()?;
+    if !path.exists() {
+        return Ok(None);
+    }
+    let bytes = fs::read(&path).with_context(|| format!("read failed: {}", path.display()))?;
+    if bytes.is_empty() {
+        return Ok(None);
+    }
+    // Stub: base64 envelope — replace with Keystore-backed AES when JNI lands.
+    use base64::Engine;
+    let decoded = base64::engine::general_purpose::STANDARD
+        .decode(bytes)
+        .context("decode android credentials envelope")?;
+    let root =
+        serde_json::from_slice::<CredsRoot>(&decoded).context("parse android credentials json")?;
+    Ok(Some(root))
+}
+
+#[cfg(target_os = "android")]
+fn save_android_credentials(root: &CredsRoot) -> Result<()> {
+    let cleaned = clean_credentials(root);
+    let json = serde_json::to_string(&cleaned).context("encode credentials failed")?;
+    use base64::Engine;
+    let encoded = base64::engine::general_purpose::STANDARD.encode(json.as_bytes());
+    let path = android_credentials_path()?;
+    super::ensure_dir(path.parent().unwrap_or_else(|| Path::new(".")))?;
+    fs::write(&path, encoded).with_context(|| format!("write failed: {}", path.display()))?;
+    Ok(())
+}
+
+fn clean_credentials(root: &CredsRoot) -> CredsRoot {
+    let mut cleaned = root.clone();
+    cleaned.providers.asr.retain(|_, v| !v.is_empty());
+    cleaned.providers.llm.retain(|_, v| !v.is_empty());
+    cleaned
+}
+
+fn read_legacy_credentials_file(path: &Path) -> Option<CredsRoot> {
+    if !path.exists() {
+        return None;
+    }
+    let bytes = match std::fs::read(path) {
+        Ok(b) => b,
+        Err(e) => {
+            log::warn!("[vault] read legacy {} failed: {}", path.display(), e);
+            return None;
+        }
+    };
+    match serde_json::from_slice::<CredsRoot>(&bytes) {
+        Ok(root) => Some(root),
+        Err(e) => {
+            log::warn!("[vault] parse legacy {} failed: {}", path.display(), e);
+            None
+        }
+    }
+}
+
+fn remove_legacy_credentials_file() -> Result<()> {
+    let Ok(path) = credentials_path() else {
+        return Ok(());
+    };
+    if path.exists() {
+        std::fs::remove_file(&path)
+            .with_context(|| format!("remove legacy credentials file {}", path.display()))?;
+    }
+    Ok(())
+}
+
+fn remove_legacy_credentials_file_best_effort() {
+    if let Err(e) = remove_legacy_credentials_file() {
+        log::warn!("[vault] remove legacy credentials file failed: {e}");
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+struct CredsChunkManifest {
+    openless_credentials_storage: String,
+    version: u32,
+    /// 旧版本（v1 早期）每次 save 都生成新 UUID 作为 chunk account 命名前缀，
+    /// 这让 macOS Keychain 的「始终允许」每次保存后失效 → 反复弹 ACL 弹窗。
+    /// 现在 save 总用稳定 chunk.{index} 名，此字段仅向后兼容旧 manifest 读取。
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    generation: Option<String>,
+    chunks: usize,
+}
+
+/// 旧版（generation=Some）：`credentials.v1.chunk.<UUID>.{index}`
+/// 新版（generation=None）：`credentials.v1.chunk.{index}` —— 稳定名，ACL 长期有效
+fn chunk_account(generation: Option<&str>, index: usize) -> String {
+    match generation {
+        Some(gen) => format!("{KEYRING_CREDENTIALS_CHUNK_PREFIX}{gen}.{index}"),
+        None => format!("{KEYRING_CREDENTIALS_CHUNK_PREFIX}{index}"),
+    }
+}
+
+fn chunk_json_payload(json: &str) -> Vec<String> {
+    let mut chunks = Vec::new();
+    let mut current = String::new();
+    let mut current_units = 0usize;
+    for ch in json.chars() {
+        let units = ch.len_utf16();
+        if !current.is_empty() && current_units + units > KEYRING_CHUNK_MAX_UTF16_UNITS {
+            chunks.push(std::mem::take(&mut current));
+            current_units = 0;
+        }
+        current.push(ch);
+        current_units += units;
+    }
+    if !current.is_empty() || json.is_empty() {
+        chunks.push(current);
+    }
+    chunks
+}
+
+fn read_chunk_manifest(json: &str) -> Option<CredsChunkManifest> {
+    let manifest = serde_json::from_str::<CredsChunkManifest>(json).ok()?;
+    if manifest.openless_credentials_storage == "chunked" && manifest.version == 1 {
+        Some(manifest)
+    } else {
+        None
+    }
+}
+
+#[cfg(not(target_os = "android"))]
+fn get_keyring_password(account: &str) -> Result<Option<String>> {
+    match keyring_entry_for(account)?.get_password() {
+        Ok(value) => Ok(Some(value)),
+        Err(keyring::Error::NoEntry) => Ok(None),
+        Err(e) => {
+            Err(anyhow!(e)).with_context(|| format!("read system credential vault {account}"))
+        }
+    }
+}
+
+#[cfg(not(target_os = "android"))]
+fn delete_keyring_password(account: &str) {
+    match keyring_entry_for(account).and_then(|entry| {
+        entry
+            .delete_credential()
+            .with_context(|| format!("delete system credential vault {account}"))
+    }) {
+        Ok(()) | Err(_) => {}
+    }
+}
+
+#[cfg(not(target_os = "android"))]
+fn load_keyring_credentials() -> Result<Option<CredsRoot>> {
+    let Some(json_or_manifest) = get_keyring_password(KEYRING_CREDENTIALS_ACCOUNT)? else {
+        return Ok(None);
+    };
+
+    let manifest = read_chunk_manifest(&json_or_manifest)
+        .ok_or_else(|| anyhow!("invalid system credential vault manifest"))?;
+    let mut json = String::new();
+    for index in 0..manifest.chunks {
+        let account = chunk_account(manifest.generation.as_deref(), index);
+        let chunk = get_keyring_password(&account)?
+            .ok_or_else(|| anyhow!("missing system credential vault chunk {index}"))?;
+        json.push_str(&chunk);
+    }
+
+    serde_json::from_str::<CredsRoot>(&json)
+        .map(Some)
+        .context("decode system credential vault payload")
+}
+
+#[cfg(not(target_os = "android"))]
+fn load_legacy_keyring_credentials() -> CredsRoot {
+    match load_legacy_keyring_credentials_for_update() {
+        Ok(root) => root,
+        Err(e) => {
+            log::warn!("[vault] read legacy vault credentials failed: {e}");
+            CredsRoot::default()
+        }
+    }
+}
+
+#[cfg(not(target_os = "android"))]
+fn load_legacy_keyring_credentials_for_update() -> Result<CredsRoot> {
+    let mut root = CredsRoot::default();
+    for account in CredentialAccount::all() {
+        let legacy_account = account.keyring_account();
+        match get_keyring_password(legacy_account) {
+            Ok(Some(value)) => write_account(&mut root, *account, Some(value)),
+            Ok(None) => {}
+            Err(e) => return Err(e.context(format!("read legacy vault {legacy_account}"))),
+        }
+    }
+    Ok(clean_credentials(&root))
+}
+
+#[cfg(not(target_os = "android"))]
+fn remove_legacy_keyring_credentials() {
+    for account in CredentialAccount::all() {
+        delete_keyring_password(account.keyring_account());
+    }
+}
+
+fn load_legacy_credentials() -> Option<CredsRoot> {
+    credentials_path()
+        .ok()
+        .and_then(|p| read_legacy_credentials_file(&p))
+}
+
+fn legacy_vault_has_credentials(root: &CredsRoot) -> bool {
+    !root.providers.asr.is_empty() || !root.providers.llm.is_empty()
+}
+
+fn load_legacy_sources_without_migration() -> CredsRoot {
+    if let Some(legacy) = load_legacy_credentials() {
+        return legacy;
+    }
+
+    #[cfg(not(target_os = "android"))]
+    {
+        let legacy_vault = load_legacy_keyring_credentials();
+        if legacy_vault_has_credentials(&legacy_vault) {
+            return legacy_vault;
+        }
+    }
+
+    CredsRoot::default()
+}
+
+fn migrate_legacy_sources() -> CredsRoot {
+    match migrate_legacy_sources_for_update() {
+        Ok(root) => root,
+        Err(e) => {
+            log::warn!("[vault] legacy credential migration failed: {e}");
+            load_legacy_sources_without_migration()
+        }
+    }
+}
+
+fn migrate_legacy_sources_for_update() -> Result<CredsRoot> {
+    if let Some(legacy) = load_legacy_credentials() {
+        save_credentials(&legacy)?;
+        #[cfg(not(target_os = "android"))]
+        remove_legacy_keyring_credentials();
+        return Ok(legacy);
+    }
+
+    #[cfg(not(target_os = "android"))]
+    {
+        let legacy_vault = load_legacy_keyring_credentials_for_update()?;
+        if legacy_vault_has_credentials(&legacy_vault) {
+            save_credentials(&legacy_vault)?;
+            remove_legacy_keyring_credentials();
+            return Ok(legacy_vault);
+        }
+    }
+
+    Ok(CredsRoot::default())
+}
+
+fn load_credentials() -> CredsRoot {
+    if let Some(cached) = credentials_cache().lock().as_ref().cloned() {
+        return cached;
+    }
+
+    #[cfg(target_os = "android")]
+    {
+        let root = match load_android_credentials() {
+            Ok(Some(root)) => root,
+            Ok(None) => CredsRoot::default(),
+            Err(e) => {
+                log::warn!("[vault] android credential read failed: {e}");
+                CredsRoot::default()
+            }
+        };
+        store_credentials_cache(&root);
+        return root;
+    }
+
+    #[cfg(not(target_os = "android"))]
+    match load_keyring_credentials() {
+        Ok(Some(root)) => {
+            // 不在这里调 remove_legacy_keyring_credentials() —— 它内部对每个
+            // 旧 account 各做一次 keyring delete，每次 delete 在 macOS Keychain
+            // 上仍要触发 ACL 检查。第一次成功 load 时 legacy entries 通常已经
+            // 被 migrate_legacy_sources_for_update 清理过了；这里若再无脑跑，
+            // 只会反复弹「OpenLess 想删除 X」十几次。文件 legacy（plaintext
+            // JSON）不需要 ACL，可继续 best-effort 删除。
+            remove_legacy_credentials_file_best_effort();
+            store_credentials_cache(&root);
+            root
+        }
+        Ok(None) => {
+            // 没有现成 chunked manifest —— 走 migrate（如果有 legacy 则写入并返回写后的 root）。
+            // migrate_legacy_sources 内部 save_credentials 已经会刷 cache，这里再补一次
+            // 是为了「无 legacy 也无 manifest」走默认 root 的路径也能进 cache。
+            let root = migrate_legacy_sources();
+            store_credentials_cache(&root);
+            root
+        }
+        Err(e) => {
+            // **不缓存 keyring 错误路径下的 fallback**。Keychain 可能只是临时不可读
+            // （用户尚未在第一次弹窗里点同意 / DataProtection 错误 / login keychain
+            // 还没 unlock）；如果在这里把 legacy fallback 写进 cache，等用户授权后
+            // 我们就再也不会重读 keyring，整个进程生命周期里都拿 stale 数据。下次
+            // 调用让它再尝试一次 keyring。pr_agent feedback on PR #394。
+            log::warn!("[vault] system credential read failed: {e}");
+            load_legacy_sources_without_migration()
+        }
+    }
+}
+
+fn load_credentials_for_update() -> Result<CredsRoot> {
+    if let Some(cached) = credentials_cache().lock().as_ref().cloned() {
+        return Ok(cached);
+    }
+
+    #[cfg(target_os = "android")]
+    {
+        let root = match load_android_credentials()? {
+            Some(root) => root,
+            None => CredsRoot::default(),
+        };
+        store_credentials_cache(&root);
+        return Ok(root);
+    }
+
+    #[cfg(not(target_os = "android"))]
+    match load_keyring_credentials() {
+        Ok(Some(root)) => {
+            // 同 load_credentials：不再每次 update 都尝试 delete legacy keyring
+            // entries，避免反复触发 macOS Keychain ACL 弹窗。
+            remove_legacy_credentials_file_best_effort();
+            store_credentials_cache(&root);
+            Ok(root)
+        }
+        Ok(None) => {
+            // migrate_legacy_sources_for_update 内部如果实际 migrate 会调
+            // save_credentials，cache 会被刷新；如果只返回 default root（没 legacy），
+            // 我们这里再显式 cache 一次防御性补一下。
+            let root = migrate_legacy_sources_for_update()?;
+            store_credentials_cache(&root);
+            Ok(root)
+        }
+        // 错误路径不缓存 —— 同 load_credentials 注释；让下次读重试 keyring。
+        Err(e) => Err(e),
+    }
+}
+
+fn save_credentials(root: &CredsRoot) -> Result<()> {
+    let cleaned = clean_credentials(root);
+
+    #[cfg(target_os = "android")]
+    {
+        save_android_credentials(&cleaned)?;
+        store_credentials_cache(&cleaned);
+        return Ok(());
+    }
+
+    #[cfg(not(target_os = "android"))]
+    {
+        let json = serde_json::to_string(&cleaned).context("encode credentials failed")?;
+        let previous_manifest = get_keyring_password(KEYRING_CREDENTIALS_ACCOUNT)
+            .ok()
+            .flatten()
+            .and_then(|value| read_chunk_manifest(&value));
+        let chunks = chunk_json_payload(&json);
+
+        // 先写所有 chunks（稳定名），再写 manifest —— 保证 partial-write 不会让
+        // manifest 指向不完整 chunks。stable name 让 macOS Keychain ACL 一次允许后
+        // 长期有效，不再因 UUID 轮换反复弹窗（这是 PR #277 早期 UUID-rotation
+        // 设计的回退）。
+        for (index, chunk) in chunks.iter().enumerate() {
+            let account = chunk_account(None, index);
+            keyring_entry_for(&account)?
+                .set_password(chunk)
+                .with_context(|| format!("write system credential vault chunk {index}"))?;
+        }
+
+        let manifest = CredsChunkManifest {
+            openless_credentials_storage: "chunked".to_string(),
+            version: 1,
+            generation: None,
+            chunks: chunks.len(),
+        };
+        let manifest_json =
+            serde_json::to_string(&manifest).context("encode credential manifest failed")?;
+        keyring_entry()?
+            .set_password(&manifest_json)
+            .context("write system credential vault manifest")?;
+
+        // 清理旧 chunks：
+        // 1) 旧 manifest 用 UUID generation → 那一代 chunks 全删（迁移到 stable name）
+        // 2) 旧 manifest 也是 stable name，但 chunks 数量比这次多 → 删多余的 idx
+        if let Some(previous) = previous_manifest {
+            match previous.generation.as_deref() {
+                Some(prev_gen) => {
+                    for index in 0..previous.chunks {
+                        delete_keyring_password(&chunk_account(Some(prev_gen), index));
+                    }
+                }
+                None => {
+                    for index in chunks.len()..previous.chunks {
+                        delete_keyring_password(&chunk_account(None, index));
+                    }
+                }
+            }
+        }
+
+        remove_legacy_credentials_file_best_effort();
+        // 写完成功后立刻刷新 process cache —— 同进程后续读不再回 Keychain。
+        // 见 CREDENTIALS_CACHE 的 doc。
+        store_credentials_cache(&cleaned);
+        Ok(())
+    }
+}
+
+fn lookup_account(root: &CredsRoot, account: CredentialAccount) -> Option<String> {
+    let asr = root.providers.asr.get(&root.active.asr);
+    let llm = root.providers.llm.get(&root.active.llm);
+    let pick = |s: &Option<String>| s.as_ref().filter(|v| !v.is_empty()).cloned();
+    match account {
+        CredentialAccount::VolcengineAppKey => {
+            asr.and_then(|e| pick(&e.appKey).or_else(|| pick(&e.apiKey)))
+        }
+        CredentialAccount::VolcengineAccessKey => asr.and_then(|e| pick(&e.accessKey)),
+        CredentialAccount::VolcengineResourceId => asr.and_then(|e| pick(&e.resourceId)),
+        CredentialAccount::ArkApiKey => llm.and_then(|e| pick(&e.apiKey)),
+        CredentialAccount::ArkModelId => llm.and_then(|e| pick(&e.model)),
+        CredentialAccount::ArkEndpoint => llm.and_then(|e| pick(&e.baseURL)),
+        CredentialAccount::AsrApiKey => asr.and_then(|e| pick(&e.apiKey)),
+        CredentialAccount::AsrEndpoint => asr.and_then(|e| pick(&e.baseURL)),
+        CredentialAccount::AsrModel => asr.and_then(|e| pick(&e.model)),
+        CredentialAccount::AsrVocabularyId => asr.and_then(|e| pick(&e.vocabularyId)),
+    }
+}
+
+fn write_account(root: &mut CredsRoot, account: CredentialAccount, value: Option<String>) {
+    let asr_id = root.active.asr.clone();
+    let llm_id = root.active.llm.clone();
+    let normalized = value.and_then(|v| if v.is_empty() { None } else { Some(v) });
+    match account {
+        CredentialAccount::VolcengineAppKey => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.appKey = normalized;
+        }
+        CredentialAccount::VolcengineAccessKey => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.accessKey = normalized;
+        }
+        CredentialAccount::VolcengineResourceId => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.resourceId = normalized;
+        }
+        CredentialAccount::ArkApiKey => {
+            let entry = root.providers.llm.entry(llm_id).or_default();
+            entry.apiKey = normalized;
+        }
+        CredentialAccount::ArkModelId => {
+            let entry = root.providers.llm.entry(llm_id).or_default();
+            entry.model = normalized;
+        }
+        CredentialAccount::ArkEndpoint => {
+            let entry = root.providers.llm.entry(llm_id).or_default();
+            entry.baseURL = normalized;
+        }
+        CredentialAccount::AsrApiKey => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.apiKey = normalized;
+        }
+        CredentialAccount::AsrEndpoint => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.baseURL = normalized;
+        }
+        CredentialAccount::AsrModel => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.model = normalized;
+        }
+        CredentialAccount::AsrVocabularyId => {
+            let entry = root.providers.asr.entry(asr_id).or_default();
+            entry.vocabularyId = normalized;
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
+pub enum CredentialAccount {
+    VolcengineAppKey,
+    VolcengineAccessKey,
+    VolcengineResourceId,
+    ArkApiKey,
+    ArkModelId,
+    ArkEndpoint,
+    /// Active ASR provider's API key (used by Whisper-compatible providers).
+    AsrApiKey,
+    /// Active ASR provider's base URL.
+    AsrEndpoint,
+    /// Active ASR provider's model name.
+    AsrModel,
+    /// Active ASR provider's optional hotword vocabulary ID.
+    AsrVocabularyId,
+}
+
+impl CredentialAccount {
+    /// Account names match the Swift `CredentialAccount` constants exactly so
+    /// existing Keychain entries written by the macOS Swift app remain
+    /// readable after upgrade.
+    pub fn keyring_account(&self) -> &'static str {
+        match self {
+            CredentialAccount::VolcengineAppKey => "volcengine.app_key",
+            CredentialAccount::VolcengineAccessKey => "volcengine.access_key",
+            CredentialAccount::VolcengineResourceId => "volcengine.resource_id",
+            CredentialAccount::ArkApiKey => "ark.api_key",
+            CredentialAccount::ArkModelId => "ark.model_id",
+            CredentialAccount::ArkEndpoint => "ark.endpoint",
+            CredentialAccount::AsrApiKey => "asr.api_key",
+            CredentialAccount::AsrEndpoint => "asr.endpoint",
+            CredentialAccount::AsrModel => "asr.model",
+            CredentialAccount::AsrVocabularyId => "asr.vocabulary_id",
+        }
+    }
+
+    pub fn all() -> &'static [CredentialAccount] {
+        &[
+            CredentialAccount::VolcengineAppKey,
+            CredentialAccount::VolcengineAccessKey,
+            CredentialAccount::VolcengineResourceId,
+            CredentialAccount::ArkApiKey,
+            CredentialAccount::ArkModelId,
+            CredentialAccount::ArkEndpoint,
+            CredentialAccount::AsrApiKey,
+            CredentialAccount::AsrEndpoint,
+            CredentialAccount::AsrModel,
+            CredentialAccount::AsrVocabularyId,
+        ]
+    }
+}
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CredentialsSnapshot {
+    pub volcengine_app_key: Option<String>,
+    pub volcengine_access_key: Option<String>,
+    pub volcengine_resource_id: Option<String>,
+    pub asr_api_key: Option<String>,
+    pub asr_endpoint: Option<String>,
+    pub asr_model: Option<String>,
+    pub ark_api_key: Option<String>,
+    pub ark_model_id: Option<String>,
+    pub ark_endpoint: Option<String>,
+}
+
+/// 凭据存储——系统凭据库；旧 JSON 文件只作为迁移来源。
+pub struct CredentialsVault;
+
+impl CredentialsVault {
+    /// 系统凭据库 service name；macOS 下对应 Keychain service。
+    pub const SERVICE_NAME: &'static str = "com.openless.app";
+
+    pub fn get(account: CredentialAccount) -> Result<Option<String>> {
+        let _guard = credentials_lock().lock();
+        Ok(lookup_account(&load_credentials(), account))
+    }
+
+    pub fn set(account: CredentialAccount, value: &str) -> Result<()> {
+        let _guard = credentials_lock().lock();
+        let mut root = load_credentials_for_update()?;
+        let v = if value.is_empty() {
+            None
+        } else {
+            Some(value.to_string())
+        };
+        write_account(&mut root, account, v);
+        save_credentials(&root)
+    }
+
+    pub fn remove(account: CredentialAccount) -> Result<()> {
+        let _guard = credentials_lock().lock();
+        let mut root = load_credentials_for_update()?;
+        write_account(&mut root, account, None);
+        save_credentials(&root)
+    }
+
+    pub fn get_active_asr() -> String {
+        let _guard = credentials_lock().lock();
+        load_credentials().active.asr
+    }
+
+    pub fn set_active_asr_provider(id: &str) -> Result<()> {
+        let _guard = credentials_lock().lock();
+        let mut root = load_credentials_for_update()?;
+        root.active.asr = id.to_string();
+        save_credentials(&root)
+    }
+
+    pub fn set_active_llm_provider(id: &str) -> Result<()> {
+        let _guard = credentials_lock().lock();
+        let mut root = load_credentials_for_update()?;
+        root.active.llm = id.to_string();
+        save_credentials(&root)
+    }
+
+    pub fn get_active_llm() -> String {
+        let _guard = credentials_lock().lock();
+        load_credentials().active.llm
+    }
+
+    pub fn snapshot() -> CredentialsSnapshot {
+        let _guard = credentials_lock().lock();
+        let root = load_credentials();
+        CredentialsSnapshot {
+            volcengine_app_key: lookup_account(&root, CredentialAccount::VolcengineAppKey),
+            volcengine_access_key: lookup_account(&root, CredentialAccount::VolcengineAccessKey),
+            volcengine_resource_id: lookup_account(&root, CredentialAccount::VolcengineResourceId),
+            asr_api_key: lookup_account(&root, CredentialAccount::AsrApiKey),
+            asr_endpoint: lookup_account(&root, CredentialAccount::AsrEndpoint),
+            asr_model: lookup_account(&root, CredentialAccount::AsrModel),
+            ark_api_key: lookup_account(&root, CredentialAccount::ArkApiKey),
+            ark_model_id: lookup_account(&root, CredentialAccount::ArkModelId),
+            ark_endpoint: lookup_account(&root, CredentialAccount::ArkEndpoint),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{chunk_json_payload, KEYRING_CHUNK_MAX_UTF16_UNITS};
+
+    #[test]
+    fn credential_payload_chunks_stay_under_windows_blob_limit() {
+        let payload = format!(
+            "{}{}{}",
+            "a".repeat(KEYRING_CHUNK_MAX_UTF16_UNITS + 25),
+            "😀".repeat(20),
+            "b".repeat(KEYRING_CHUNK_MAX_UTF16_UNITS + 25)
+        );
+        let chunks = chunk_json_payload(&payload);
+        assert!(chunks.len() > 1);
+        assert_eq!(chunks.concat(), payload);
+        assert!(chunks
+            .iter()
+            .all(|chunk| chunk.encode_utf16().count() <= KEYRING_CHUNK_MAX_UTF16_UNITS));
+    }
+}
